I converted the entire code towards one array and storing all information about tasks in objects which were stored in said array.
This reduced the need to add/subtract elements from 2-3 arrays and reduced the complexity of the code.

The original version of the localStorage exercises used an array as a kind of reference file for the separate "files" stored in the localStorage... which led to its own kind of complexity.

Storing status and changes as object properties was much simpler as I only had to pass one object to a function and access its properties...  as opposed to having to enter multiple different arguments for different situations.

Now I could use conditionals based on object property values to write the code.